+++ app/frontend/src/vite-env.d.ts
@@ -1,1 +1,5 @@
 /// <reference types="vite/client" />
+
+declare module '*/package.json' {
+  export const version: string
+}